DSS-2152 Save Commingle in synoptical and voyage status
* DSS-2173 Java -   Enhance get synoptical table API
Type : Feature

* DSS-2173

* Post synoptical table api enhancement
Type: Feature

* Post synoptical table api enhancement

* post synoptical table

* DSS-2175 Save loadicator info enhancement

* DSS-2173 Added actualQuantity to loadablePlanComminglePortwiseDetails entity

* DSS-2173 -set actual quantity in gateway

* DSS-2172
Java -Modify the logic for showing cargo sum
Type :feature

* DSS-2172 Java -Modify the logic for showing cargo sum .
Type :feature

DSS-1437
Type :Bug

* loadicator info

* merge issues

+++ api/cp-dss-api/loadable-study/src/main/java/com/cpdss/loadablestudy/entity/LoadablePlanComminglePortwiseDetails.java
@@ -2,6 +2,7 @@
 package com.cpdss.loadablestudy.entity;
 
 import com.cpdss.common.utils.EntityDoc;
+import java.math.BigDecimal;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.JoinColumn;
@@ -126,5 +127,5 @@ public class LoadablePlanComminglePortwiseDetails extends EntityDoc {
   private Long portRotationXid;
 
   @Column(name = "actual_quantity")
-  private Long actualQuantity;
+  private BigDecimal actualQuantity;
 }
